refactor: restructure `tycho_storage`
