+++ Cargo.toml
@@ -50,7 +50,7 @@ dirs = "5.0.1"
 ed25519 = "2.0"
 ed25519-dalek = "2.1"
 everscale-crypto = { version = "0.3", features = ["tl-proto", "serde"] }
-everscale-types = { version = "0.1.2", features = ["tycho"] }
+everscale-types = { version = "0.1.2", features = ["tycho", "stats"] }
 exponential-backoff = "1"
 fdlimit = "0.3.0"
 futures-util = "0.3"
