+++ cli/Cargo.toml
@@ -30,7 +30,6 @@ futures-util = { workspace = true }
 hex = { workspace = true }
 humantime = { workspace = true }
 tabled = { workspace = true, default-features = false, features = ["std"] }
-metrics = { workspace = true, optional = true }
 parking_lot = { workspace = true }
 rand = { workspace = true }
 rayon = { workspace = true }
@@ -64,7 +63,6 @@ tycho-consensus = { workspace = true }
 
 [dev-dependencies]
 tycho-collator = { workspace = true, features = ["test"] }
-tycho-storage = { workspace = true, features = ["test"] }
 num-bigint = { workspace = true }
 
 [build-dependencies]
@@ -73,7 +71,7 @@ rustc_version = { workspace = true }
 
 [features]
 default = ["jemalloc"]
-jemalloc = ["dep:tikv-jemallocator", "dep:tikv-jemalloc-ctl", "dep:metrics"]
+jemalloc = ["dep:tikv-jemallocator", "dep:tikv-jemalloc-ctl"]
 deadlock-detection = ["parking_lot/deadlock_detection"]
 debug = ["tycho-consensus/test"]
 lto = ["weedb/lto"]
