+++ collator/src/storage/db.rs
@@ -1,23 +1,15 @@
-use tycho_storage::{Migrations, WithMigrations};
-use tycho_util::sync::CancellationFlag;
-use weedb::{Caches, MigrationError, Semver, WeeDb};
+use tycho_storage::kv::NamedTables;
+use weedb::{Caches, WeeDb};
 
 use super::tables;
 
 pub type InternalQueueDB = WeeDb<InternalQueueTables>;
 
-impl WithMigrations for InternalQueueTables {
+impl NamedTables for InternalQueueTables {
     const NAME: &'static str = "int_queue";
-    const VERSION: Semver = [0, 0, 1];
-
-    fn register_migrations(
-        _migrations: &mut Migrations<Self>,
-        _cancelled: CancellationFlag,
-    ) -> Result<(), MigrationError> {
-        // TODO: register migrations here
-        Ok(())
-    }
 }
+
+// TODO: Add migrations.
 
 weedb::tables! {
     pub struct InternalQueueTables<Caches> {
