+++ core/Cargo.toml
@@ -54,7 +54,6 @@ tempfile = { workspace = true }
 tycho-block-util = { workspace = true, features = ["test"] }
 tycho-core = { workspace = true, features = ["test"] }
 tycho-network = { workspace = true, features = ["test"] }
-tycho-storage = { workspace = true, features = ["test"] }
 tycho-util = { workspace = true, features = ["test"] }
 
 [features]
