+++ core/src/storage/shard_state/store_state_raw.rs
@@ -7,7 +7,8 @@ use everscale_types::cell::*;
 use everscale_types::models::BlockId;
 use everscale_types::util::ArrayVec;
 use tycho_block_util::state::*;
-use tycho_storage::{MappedFile, StoredValue, TempFileStorage};
+use tycho_storage::fs::{MappedFile, TempFileStorage};
+use tycho_storage::kv::StoredValue;
 use tycho_util::io::ByteOrderRead;
 use tycho_util::progress_bar::*;
 use tycho_util::FastHashMap;
