--- storage/src/db/mod.rs
@@ -1,5 +1,0 @@
-pub use self::file_db::*;
-pub use self::kv_db::*;
-
-mod file_db;
-mod kv_db;
