+++ storage/src/lib.rs
@@ -1,17 +1,7 @@
-pub use tycho_storage_traits::{StoredValue, StoredValueBuffer};
-
 pub use self::config::*;
 pub use self::context::StorageContext;
-pub use self::db::*;
-pub use self::store::*;
 
 mod config;
 mod context;
-mod db;
-mod store;
-
-pub mod util {
-    pub use self::instance_id::*;
-
-    pub mod instance_id;
-}
+pub mod fs;
+pub mod kv;
