--- storage/src/store/mod.rs
@@ -1,3 +1,0 @@
-pub use self::temp_file::*;
-
-mod temp_file;
